pyproject.toml
[project]
name = "safie-tmp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "anthropic>=0.86.0",
    "boto3>=1.42.27",
    "google-auth>=2.40.3",
    "google-cloud-storage>=3.4.1",
    "google-genai>=1.17.0",
    "marker-pdf>=1.6.2",
    "pdfplumber>=0.11.9",
    "pillow>=10.1.0,<11.0.0",
    "pycocotools>=2.0.10",
    "pydantic>=2.11.9",
    "pymupdf4llm>=1.27.2.1",
    "python-dotenv>=1.1.1",
]

[tool.ruff]
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "standard"

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "ruff>=0.15.7",
]

[tool.uv]
override-dependencies = [
    "anthropic>=0.86.0",
    "regex>=2024.4.28",
]
